refactor: move command palette catalog into app
