+++ src/app/mod.rs
@@ -1,6 +1,7 @@
 pub mod cmd;
 pub mod input;
 pub mod model;
+pub mod palette;
 pub mod policy;
 pub mod runtime;
 pub mod startup;
@@ -15,6 +16,7 @@ pub mod app {
     pub use crate::cmd;
     pub use crate::input;
     pub use crate::model;
+    pub use crate::palette;
     pub use crate::policy;
     pub use crate::ports;
     pub use crate::runtime;
