+++ src/app/palette.rs
@@ -1,5 +1,5 @@
-use super::keybindings::{GLOBAL_KEYS, KeyBinding, idx};
-use crate::app::update::action::Action;
+use crate::update::action::Action;
+use crate::update::input::keybindings::{GLOBAL_KEYS, KeyBinding, idx};
 
 // - PALETTE: opening the palette from inside itself makes no sense
 // - COMMAND_LINE: command-line mode is a separate entry mechanism
