+++ src/app/update/browse/navigation/input.rs
@@ -2,7 +2,7 @@ use crate::app::cmd::effect::Effect;
 use crate::app::model::app_state::AppState;
 use crate::app::model::shared::input_mode::InputMode;
 use crate::app::update::action::{Action, InputTarget, ListMotion, ListTarget};
-use crate::app::update::input::palette::palette_command_count;
+use crate::app::palette::palette_command_count;
 
 pub fn reduce(state: &mut AppState, action: &Action) -> Option<Vec<Effect>> {
     match action {
