+++ src/app/update/input/mod.rs
@@ -3,10 +3,8 @@
 //! - `keybindings`: surface-specific key handling
 //! - `keymap`: low-level key translation
 //! - `command`: command-line input flow
-//! - `palette`: command palette-specific input helpers
 
 pub mod command;
 pub mod keybindings;
 pub mod keymap;
-pub mod palette;
 pub mod vim;
